requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=production.txt production.in
#
aiofiles==23.1.0
    # via -r production.in
aiologger==0.7.0
    # via -r production.in
aiosmtpd==1.4.4.post2
    # via -r production.in
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via httpcore
asgiref==3.7.2
    # via django
atpublic==4.0
    # via aiosmtpd
attrs==23.1.0
    # via aiosmtpd
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   influxdb-client
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   uvicorn
    #   wait-for-it
cryptography==39.0.2
    # via
    #   fido2
    #   python-u2flib-server
django==4.2.4
    # via
    #   -r production.in
    #   django-anymail
    #   django-extensions
    #   django-filter
    #   django-guardian
    #   django-mfa2
    #   djangorestframework
    #   drf-nested-routers
django-anymail==10.1
    # via -r production.in
django-environ==0.10.0
    # via -r production.in
django-extensions==3.2.3
    # via -r production.in
django-filter==23.2
    # via -r production.in
django-guardian==2.4.0
    # via -r production.in
django-mfa2==2.8.0
    # via -r production.in
django-ordered-model==3.7.4
    # via -r production.in
djangorestframework==3.14.0
    # via
    #   -r production.in
    #   drf-nested-routers
dnspython==2.4.1
    # via -r production.in
drf-nested-routers==0.93.4
    # via -r production.in
ecdsa==0.18.0
    # via python-jose
exceptiongroup==1.1.2
    # via anyio
fido2==1.0.0
    # via django-mfa2
gunicorn==21.2.0
    # via -r production.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via -r production.in
huey==2.4.5
    # via -r production.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
influxdb-client==1.37.0
    # via -r production.in
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==23.1
    # via gunicorn
postmarker==1.0
    # via -r production.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==2.1.1
    # via -r production.in
pydantic-core==2.4.0
    # via pydantic
pygments==2.15.1
    # via rich
pyotp==2.9.0
    # via django-mfa2
python-dateutil==2.8.2
    # via influxdb-client
python-jose==3.3.0
    # via django-mfa2
python-u2flib-server==5.0.1
    # via django-mfa2
pytz==2023.3
    # via
    #   -r production.in
    #   djangorestframework
pyyaml==6.0.1
    # via -r production.in
reactivex==4.0.4
    # via influxdb-client
redispy==3.0.0
    # via -r production.in
requests==2.31.0
    # via
    #   django-anymail
    #   postmarker
rich==13.5.2
    # via -r production.in
rsa==4.9
    # via python-jose
simplejson==3.19.1
    # via django-mfa2
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
    #   python-u2flib-server
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via
    #   asgiref
    #   pydantic
    #   pydantic-core
    #   reactivex
    #   uvicorn
ua-parser==0.18.0
    # via
    #   django-mfa2
    #   user-agents
urllib3==2.0.4
    # via
    #   django-anymail
    #   influxdb-client
    #   requests
user-agents==2.2.0
    # via django-mfa2
uvicorn==0.23.2
    # via -r production.in
wait-for-it==2.2.2
    # via -r production.in
whitenoise==6.5.0
    # via -r production.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
